Agrega codigo realizado en taller de Vistas basadas en Clases
