+++ paulwalker/templates/base.html
@@ -10,9 +10,6 @@
     <!-- Optional theme -->
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.4/css/bootstrap-theme.min.css">
 
-    <!-- Latest compiled and minified JavaScript -->
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.4/js/bootstrap.min.js"></script>
-
 
 </head>
 <body>
@@ -31,18 +28,25 @@
         <div id="navbar" class="collapse navbar-collapse">
           <ul class="nav navbar-nav">
             <li class="active"><a href="#">Home</a></li>
-            <li><a href="#about">About</a></li>
-            <li><a href="#contact">Contact</a></li>
+            <li><a href="{% url 'listar_marcas' %}">Marcas</a></li>
+            <li><a href="{% url 'listar_carros' %}">Carros</a></li>
           </ul>
         </div><!--/.nav-collapse -->
       </div>
     </nav>
+    <br><br>
+    <div class="navbar">
+          {% if messages %}
+            {% for message in messages %}
+                <div class="alert alert-{{ message.tags }} alert-dismissible fade in" role="alert">
+                    <button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button>
+                    <p>{{ message }}</p>
+                </div>
+            {% endfor %}
+          {% endif %}
+        </div>
 
     <div class="container">
-        <br>
-        <br>
-        <br>
-        <br>
       <div class="starter-template">
           {% block content %}
             <h1>Hola mundo desde templates</h1>
@@ -52,6 +56,8 @@
 
     </div><!-- /.container -->
 
-
+    <!-- Latest compiled and minified JavaScript -->
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.3/jquery.min.js"></script>
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.4/js/bootstrap.min.js"></script>
 </body>
 </html>
